day 2
The task wasn't too challenging today but I'm still having
troubles with Haskell. First I don't fully understand the rules
around function composition such as when are you allowed to use
`.` vs `$`. Second there seems to be some annoying restrictions on
return types, for instance it seems that returning a list of lists
of tuples is not valid but returning a tuple of lists of tuples
is.

Again this was too fast to get repeatable performance numbers.
Even more so than yesterday. Execution time was on the order of
150us.