day3 c solution
This day was pretty rough, mostly due to the fact that the
provided test data missed most edge cases. The one I provided
should work better. For me the main thing hanging me up was when
there was 'N\nN'. Everything else was straight forward.

I worked on the Haskell solution for most of the day but it was a
complete slog. I find that Haskell is really unfriendly to
debugging. There is no way to print some intermediate value in the
middle of a function. So far this has to be the biggest negative
for me w.r.t. Haskell. I got frustrated and implemented it in C in
a fraction of the time. With this solution in hand it shouldn't be
difficult to fix my Haskell version but it will come later.

Runtime for this C version is very fast <550us. I will see if the
Haskell version performs worse or not.