day 4
This was a walk in the park compared to the hellscape of day 3.
Part 1 was a 10 minute solution but I spent some more time of Part
2 to avoid reparsing the input over and over.

The solution I came up with was to store pairs of:

(Appearances of card # N, # of winning numbers for card N)

Then I recursively walked through this list incrementing the tail
based on the number of winning numbers in the head multiplied by
the number of times the head element appeared. This meant I needed
a single recursive step per each card with its input 1 element
shorter than the previous step. This made everything very fast
(runtime on the order of 650us).

I made good use of function composition in this one. The only ugly
function is the recursive one because I couldn't come up with a
good way of incrementing the remaining cards.