day 8
Day 8 started out as a nice problem with some interesting
solutions but quickly grew frustrating with part 2 (partly because
of Haskell and partly because of the problem itself).

To start with I immediately compressed the input by ~33% by
converting all the nodes to base 26 integers (3 ASCII digits
require at least 24 bits, but converting to base 26 requires 26^3
values which fits in 16 bits). The instructions were converted
into a list of functions the head of which is applied to the
current node to get to the next node. Then the path is easily
traversed.

Part 2 is also fairly trivial to implement the, correct, general
solution but the inputs were chosen such that the runtime would be
much too long (apparently some people did it in reasonable runtime
using more efficient languages). As it turns out however, the
inputs were also chosen such that you could just find the LCM of
all the paths. Its just that this was not indicated anywhere (I
had originally considered this approach but discarded it before
writing because it would not be correct in general). After
struggling for a while not knowing if my general algorithm was
correct (it is) I learned that people were just using LCM. I would
have found out this solution would work sooner if Haskell had any
reasonable way of printing out intermediate values because I would
have noticed the cyclic nature of each path. But alas my long
lived frustration with this aspect of Haskell continues.

Even using the LCM approach the solution is relatively slow. I
reckon I could write a much faster version in C but I'm not sure I
care enough about this specific problem to do so. Runtime:

avgtime -n 64 ./day8 input
real 1.390469
user 1.372344
sys 0.000469

(I realized that my avgtime script doesn't play nice with data
from stdin. I will fix later or replace with a simple C program)